requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.0.300
openai>=1.0.0
redis>=4.6.0
celery>=5.3.0
pytest>=7.3.1
networkx>=3.1
psycopg2-binary>=2.9.6
python-jose[cryptography]>=3.3.0
prometheus-client>=0.17.0